+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.33.0
 * Fix error when clicking on new idea button in mobile view
+* Make votebuttons in idea-details configurable
 
 ## 0.32.0
 * Add new configuration options to reactions